[09] Sensor Boost: refactor that allows relative mode for destination argument
The `get` lambda in src/intcode.rs did not have an equivalent `set`.
It was not possible to write this lambda, since the `get` one took
a mutable borrow on `m`, the program memory.

By rewriting the run function to use a struct (`VM`), it was possible
to change `get` to a function `arg` that returns a reference to the
desired memory position, allowing it to be used as either LHS or RHS
of an assignment statement, even better than `get`/`set` functions!