[14] Space Stoichiometry: solve part one
The first tricky part was to think out the recursive algorithm for this,
which has a kind of "cheat" which is keeping state in the `surplus`
field of the `Reactor` struct. You could keep it pure and do everything
via return values and accumulating lists and whatnot, but that's much
more tedious.

The other tricky part was the `Reactor` struct, which contains a
`HashMap` that has references into the `Vec<Reaction>` that's the result
of the parsing phase. I had to indicate a lifetime (`'a`) on the
`Reactor` struct, which is also used for the key and value types. Then
the function that transforms the `Vec` into a `HashMap` has to refer to
that same lifetime, and I had to find out the syntax `impl<'a>
Reactor<'a>` to tell Rust that I want to bind the struct lifetime `'a`
to the impl lifetime `'a`.

I did the research in a [Rust playground][playground], see also the
[gist][gist].

[playground]: https://play.rust-lang.org/?version=stable&mode=release&edition=2018&gist=1beb3c22ed1c741f94834cf5e6770fa2
[gist]: https://gist.github.com/j0057/d727068ba2b607819de41277f19eb185